Разделяет store.
