+++ src/store/main/api-actions.js
@@ -1,0 +1,7 @@
+import {ActionCreator} from './action';
+import {APIRoutes} from '../../const';
+
+export const fetchOffers = () => (dispatch, _getState, api) => (
+  api.get(APIRoutes.OFFERS)
+    .then(({data}) => dispatch(ActionCreator.loadOffers(data)))
+);
