+++ src/store/middlewares/redirect.js
@@ -1,4 +1,4 @@
-import {ActionType} from '../action';
+import {ActionType} from './action';
 import browserHistory from '../../browser-history';
 
 export const redirect = (_store) => (next) => (action) => {
