+++ src/store/user/reducer.js
@@ -1,0 +1,26 @@
+import {ActionType} from './action';
+import {AuthorizationStatus} from '../../const';
+
+const initialState = {
+  authorizationStatus: AuthorizationStatus.NO_AUTH,
+  authInfo: {},
+};
+
+const reducer = (state = initialState, action) => {
+  switch (action.type) {
+    case ActionType.REQUIRED_AUTHORIZATION:
+      return {
+        ...state,
+        authorizationStatus: action.payload
+      };
+    case ActionType.SET_AUTH_INFO:
+      return {
+        ...state,
+        authInfo: action.payload
+      };
+    default:
+      return state;
+  }
+};
+
+export {reducer};
